+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Dropped support for .NET Core 3.1
+- Added support for .NET 7
 
 ## [0.4.0](https://github.com/xt0rted/tailwindcss-tag-helpers/compare/v0.3.0...v0.4.0) - 2022-10-13