+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [0.6.0](https://github.com/xt0rted/tailwindcss-tag-helpers/compare/v0.5.0...v0.6.0) - 2024-04-09
+
 - Dropped support for .NET 6
 - Dropped support for .NET 7
 - Added support for .NET 8